requirements.txt
openwakeword>=0.4.0
numpy>=1.20.0
sounddevice>=0.4.5
twilio>=8.0.0
python-dotenv>=1.0.0
onnxruntime>=1.14.0
tensorflow-lite>=2.12.0
pyaudio>=0.2.13
wave>=0.0.2
python-logging>=0.4.9.6